requirements.txt
psycopg2-binary==2.9.1
SQLAlchemy==1.4.31
python-dotenv==0.21.0
python-telegram-bot==13.14
alembic==1.8.1